refactor(api): MFW 资源端点并入 scripts 路由，不单开模块
/maafw/preview、/maafw/update 早就在 app/api/scripts.py 里，为一个图片端点另起
app/api/maafw.py 会让同一个域出现两个命名空间（/api/scripts/maafw/* 与
/api/maafw/*）和两套 OpenAPI tag（MaaFW 与 MFW）。

端点连同 _maafw_asset_file_path 与后缀白名单整体搬进 scripts.py，路径改为
/api/scripts/maafw/asset，前端 buildMaaFWAssetUrl 同步；删掉独立模块与它在
app/api/__init__.py、main.py 的注册。安全边界不变。

顺带修掉两处 prettier 报错，并清掉 useMaaFWApi 里「尚未迁移资源端点」的过期注释。
